Backend -- contrat API des 6 decisions produit (series, formats, achat, primaire, PAL)
Les nouveaux champs du schema v2 sont exposes par l'API :
- BookOut/BookCreate/BookUpdate : series (nom upsert), series_index,
  formats (possession par format, remplacement complet), price_paid,
  purchased_at, is_primary_reading (bool), tbr_rank, tbr_note
- GET /series + GET /series/{id}/books (filtre Serie, tri par tome)
- GET /books?sort=tbr_rank (PAL ordonnee, livres sans rang en fin)

Regles metier :
- primaire exclusif : PATCH is_primary_reading=true deset l'actuel dans la
  meme transaction (index partiel unique en filet), refuse hors status
  reading (422)
- wishlist sans prix : ecrire price_paid/purchased_at sur un wishlist ->
  422 sur tous les chemins (refus explicite, jamais de vidage silencieux)
- tbr_rank derive du statut : force a NULL hors de la PAL

Fixes au passage :
- series_index type Decimal cote modele (ce que renvoie la colonne NUMERIC
  de SQLite), converti en float au contrat API — elimine un warning de
  serialisation Pydantic traque jusqu'au model_dump() du modele

171 tests pytest (145 + 26 nouveaux sur formats/series/prix/primaire/PAL).

+++ backend/app/models.py
@@ -10,6 +10,7 @@ défaut d'AutoString, pour rester conforme à la lettre du DDL.
 """
 
 from datetime import datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import CheckConstraint, Index, Numeric, Text, UniqueConstraint
 # Importé sous un autre nom : le champ `text` de Highlight masquerait
@@ -75,10 +76,12 @@ class Book(SQLModel, table=True):
     acquired_date: str | None = Field(default=None, sa_type=Text)
     # Série (décision produit 15/08) : `series_index` autorise les décimales
     # (1.5 pour un hors-série). NULL = le livre n'appartient à aucune série.
+    # Typé Decimal : c'est ce que renvoie la colonne NUMERIC de SQLite (et
+    # `model_dump()` d'un champ float serait incohérent avec la valeur).
     series_id: int | None = Field(
         default=None, foreign_key="series.id", ondelete="SET NULL"
     )
-    series_index: float | None = Field(default=None, sa_type=Numeric(5, 2))
+    series_index: Decimal | None = Field(default=None, sa_type=Numeric(5, 2))
     # Prix payé / date d'achat : un seul champ chacun, jamais remplis pour un
     # livre en wishlist (le prix y serait « constaté », pas « payé »).
     price_paid: float | None = None
